Problem Statement 1
The overall goal of this project is to help the CMs focus on areas that will foster economic development for their respective states. Since the most common measure of economic development is the GDP, you will analyze the GDP of the various states of India and suggest ways to improve it.